Renamed and refactored components
