--- src/API Calls
@@ -1,18 +1,0 @@
-
-
-function getAPIData(info) {
-  const fetchedInfo = fetch(`https://rancid-tomatillos.herokuapp.com/api/v2/${info}`)
-  .then((res) => res.json())
-  return fetchedInfo
-}
-
-const getAllData = () => {
-  Promise.all([getAPIData('movies'), getAPIData('movie_id')])
-  .then((data) => {
-    movies = data[0]
-    singleMovie = data[1]
-  })
-  .catch(err => console.log("Error", err))
-}
-
-export default { getAllData }
