+++ src/components/Movie/Movie.js
@@ -1,35 +1,16 @@
-// import React from "react";
-// import Card from '../Card/Card'
-// import './Movie.css'
+import React from "react";
+import App from "../../App";
+import './Movie.css'
 
-// const Movie = ({singleMovie, showSingleMovie}) => {
-//   console.log(singleMovie)
-//   const movieCard = singleMovie.map(movie => {
-//     return (
-//       <Card
-//         id={movie.id}
-//         key={movie.id}
-//         title={movie.title}
-//         posterPath={movie.poster_path}
-//         backdropPath={movie.backdrop_path}
-//         averageRating={movie.averageRating}
-//         releaseDate={movie.release_date}
-//         overview={movie.overview}
-//         genres={movie.genres}
-//         budget={movie.budget}
-//         revenue={movie.revenue}
-//         runtime={movie.runtime}
-//         tagline={movie.tagline}
-//         showSingleMovie={showSingleMovie}
-//       />
-//     )
-//   })
+const Card = ({id, posterPath, backdropPath, title, averageRating, releaseDate, showSingleMovie}) => {
+  return (
+    <div onClick={() => showSingleMovie(id)} className="Card"> 
+      <img src={posterPath} alt="Movie Poster" width="300" height="520"></img>
+      <h2>{title}</h2>
+      <p>{averageRating}</p>
+      <p>{releaseDate}</p>
+    </div>
+  )
+}
 
-//   return (
-//     <div className="movie-container">
-//       {movieCard}
-//     </div>
-//   )
-// }
-
-// export default Movie
+export default Card
